+++ src/qqbotsdk/queue.py
@@ -13,13 +13,13 @@ class EventQueue:
         self.reply_queue: Queue[Payload] = Queue()
 
     async def put_event(self, item: Payload) -> None:
-        return await self.msg_queue.put(item)
+        await self.msg_queue.put(item)
 
     async def get_event(self) -> Payload:
         return await self.msg_queue.get()
 
     async def put_reply(self, item: Payload) -> None:
-        return await self.reply_queue.put(item)
+        await self.reply_queue.put(item)
 
     async def get_reply(self) -> Payload:
         return await self.reply_queue.get()